react-ui: rebuild the step chain as a Stepper, in its own directory
The chain becomes `Stepper`, and it draws a fixed plan the way a stepper does:
circles joined by lines that flex, so the stages spread across whatever width
they are given and the gaps stay even at three, five or eight. Measured in
Chromium at 616px: 135x4, 286x2, 71x7.

Two readings on one drawing. Which stage is in flight says where the run is in
its plan; the fill of the line leaving that stage says how far through it. A
stage with nothing to count cannot be drawn that way, so it spins instead.

That removes the windowing the old chain needed — every stage is drawn, so
nothing is measured and nothing animates its width. `motion` and
`react-resize-detector` go with it: react-ui gained both for this component two
commits ago and needs neither now.

`ProgressBar` becomes `Progress`, the bar and nothing else. The readout that
composes the two keeps its own name, `ProgressMeter`, and its own directory —
`ProgressProps` can only mean one thing, and the bar has the better claim.
`ProgressState` moves with the meter, so `Progress` and `Stepper` are both free
of it.

Three behaviours the stories exposed:

- A rewind snaps rather than eases. Sliding the fill backwards animates a run
  that never happened, and a reset taking half a second to land reads as
  still-running.
- An uncounted run that fails fills the whole bar red. It cannot say how far it
  got, so what it reports is the failure, not a fraction — and it stops sweeping,
  since motion after the run is over reads as still working.
- `TextCrawl` scrolls again under `greedy`. `greedy` holds the index at the tail
  for the whole run, and the animation was suppressed whenever the index WAS the
  tail — so it was suppressed forever. It now means arrive at the last line on
  the first paint only.
