+++ packages/sdk/shell/src/panels/StatusPanel/StatusPanel.tsx
@@ -15,7 +15,7 @@ export const StatusPanel = ({ titleId }: { titleId?: string }) => {
       <p id={titleId} className='font-medium text-center'>
         {t('resetting.message')}
       </p>
-      <Progress.Bar indeterminate>{t('resetting.message')}</Progress.Bar>
+      <Progress indeterminate>{t('resetting.message')}</Progress>
     </div>
   );
 };
