+++ packages/ui/react-ui/package.json
@@ -83,12 +83,10 @@
     "error-stack-parser": "catalog:",
     "i18next": "catalog:",
     "keyborg": "catalog:",
-    "motion": "catalog:",
     "react-aria-components": "catalog:",
     "react-error-boundary": "catalog:",
     "react-i18next": "catalog:",
-    "react-remove-scroll": "catalog:",
-    "react-resize-detector": "catalog:"
+    "react-remove-scroll": "catalog:"
   },
   "devDependencies": {
     "@dxos/random": "workspace:*",
