+++ packages/ui/react-ui/src/components/Progress/index.ts
@@ -3,6 +3,3 @@
 //
 
 export * from './Progress';
-export * from './ProgressSteps';
-export * from './Steps';
-export * from './types';
