+++ packages/ui/react-ui/src/components/ProgressMeter/index.ts
@@ -1,0 +1,6 @@
+//
+// Copyright 2026 DXOS.org
+//
+
+export * from './ProgressMeter';
+export * from './types';
