+++ packages/ui/react-ui/src/components/Stepper/index.ts
@@ -1,0 +1,5 @@
+//
+// Copyright 2026 DXOS.org
+//
+
+export * from './Stepper';
