+++ packages/ui/react-ui/src/components/index.ts
@@ -26,6 +26,7 @@ export * from './Banner';
 export * from './Panel';
 export * from './Popover';
 export * from './Progress';
+export * from './ProgressMeter';
 export * from './ScrollArea';
 export * from './ScrollContainer';
 export * from './Select';
@@ -33,6 +34,7 @@ export * from './Separator';
 export * from './Skeleton';
 export * from './Slider';
 export * from './Splitter';
+export * from './Stepper';
 export * from './Tag';
 export * from './TextCrawl';
 export * from './Toast';
